Expronicon

Collective tools for metaprogramming on Julia Expr, the meta programming standard library for MLStyle.

Meta programming in general can be decomposed into three steps:

analyse the given expression

transform the given expression to:

another target expression

a convenient intermediate representation (IR) that is easy to further manipulate/analyse

generate the target code, the target code can be:

most commonly the Julia Expr

another language's abstract/concrete syntax tree

a lower-level IR

The package Expronicon is thus written to assist you writing meta programs in the above manner.

Builtin Syntax Types

One of the most important tool in Expronicon is the syntax types, these are types describing a specific Julia syntax, e.g JLFunction describes Julia's function definition syntax. The syntax type uses a canonical data structure to represent various syntax that has the same semantic, which is convenient when one wants to manipulate, generate such objects.

They will allow you to:

easily analysis a given Julia Expr by converting it to the syntax type

easily manipulate a given Julia Expr with data structure designed for easier manipulation

easily generate a Julia Expr using codegen_ast.

Let's take JLFunction as our example again, in Julia, function can be declared in many different ways:

You can define a function using the short form

f(x) = x

or you can declare the same function using the function keyword

function f(x)
    return x
end

If we look at their expression object, we will find they actually have quite different expression structure:

ex1 = @expr f(x) = x

:(f(x) = begin
          #= none:1 =#
          x
      end)

ex2 = @expr function f(x)
    return x
end

:(function f(x)
      #= none:1 =#
      #= none:2 =#
      return x
  end)

here we use a convenient tool to obtain the Julia expression object provided by Expronicon, the @expr macro.

Now if we convert them to the JLFunction type

jl1 = JLFunction(ex1)

f(x) = begin
    #= none:1 =#    
    x    
end

jl2 = JLFunction(ex2)

function f(x)
    #= none:1 =#    
    #= none:2 =#    
    return x    
end

we can see they have the same structure under the representation of JLFunction.

dump(jl1)

JLFunction
  head: Symbol =
  name: Symbol f
  args: Array{Any}((1,))
    1: Symbol x
  kwargs: Nothing nothing
  whereparams: Nothing nothing
  body: Expr
    head: Symbol block
    args: Array{Any}((2,))
      1: LineNumberNode
        line: Int64 1
        file: Symbol none
      2: Symbol x
  line: Nothing nothing
  doc: Nothing nothing

dump(jl2)

JLFunction
  head: Symbol function
  name: Symbol f
  args: Array{Any}((1,))
    1: Symbol x
  kwargs: Nothing nothing
  whereparams: Nothing nothing
  body: Expr
    head: Symbol block
    args: Array{Any}((3,))
      1: LineNumberNode
        line: Int64 1
        file: Symbol none
      2: LineNumberNode
        line: Int64 2
        file: Symbol none
      3: Expr
        head: Symbol return
        args: Array{Any}((1,))
          1: Symbol x
  line: Nothing nothing
  doc: Nothing nothing

we can easily access to some important information of this function by accessing the fields

julia> jl1.name
:f

julia> jl1.args
1-element Vector{Any}:
 :x

julia> jl1.body
quote
    #= none:1 =#
    x
end

This is the same for other syntax types, e.g we can get the corresponding syntax type instance of a struct definition

def = @expr JLStruct struct Foo{T} <: AbstractType
    x::Int
    y::T
end

struct Foo{T} <: AbstractType
    #= none:2 =#
    x::Int
    #= none:3 =#
    y::T
end

we again use @expr for convenience, however you can also just convert the expression to JLStruct manually

ex = quote
    struct Foo{T} <: AbstractType
        x::Int
        y::T
    end
end
def = JLStruct(ex.args[2])

struct Foo{T} <: AbstractType
    #= none:3 =#
    x::Int
    #= none:4 =#
    y::T
end

once you have the corresponding JLStruct object, you can access many useful information directly

julia> def.name
:Foo

julia> def.typevars
1-element Vector{Any}:
 :T

julia> def.supertype
:AbstractType

julia> typeof(def.fields[1])
JLField

julia> def.fields[1].name
:x

julia> def.fields[1].type
:Int

Some syntax types are defined for easy manipulation such as JLIfElse, Julia's representation of if ... elseif ... else ... end statement is a recursive tree in Expr, which sometimes is not very convenient to manipulate or analysis, for example, it is not easy to access all the conditions in a long ifelse statement

ex = @expr if x > 100
    x + 1
elseif x > 90
    x + 2
elseif x > 80
    x + 3
else
    error("some error msg")
end

:(if x > 100
      #= none:2 =#
      x + 1
  elseif #= none:3 =# x > 90
      #= none:4 =#
      x + 2
  elseif #= none:5 =# x > 80
      #= none:6 =#
      x + 3
  else
      #= none:8 =#
      error("some error msg")
  end)

we can find each condition

julia> ex.args[1]
:(x > 100)

julia> ex.args[3].args[1]
quote
    #= none:3 =#
    x > 90
end

julia> ex.args[3].args[3].args[1]
quote
    #= none:5 =#
    x > 80
end

julia> ex.args[3].args[3].args[3]
quote
    #= none:8 =#
    error("some error msg")
end

imagine how would you construct such expression from scratch, or how would you access all the conditions. Thus JLIfElse allows you to access/manipulate ifelse statements directly as a dict-like object

julia> jl = JLIfElse(ex);

julia> jl.map
OrderedCollections.OrderedDict{Any, Any} with 3 entries:
  :(x > 100)                            => quote…
  quote…                                => quote…
  quote…                                => quote…

julia> jl.otherwise
quote
    #= none:8 =#
    error("some error msg")
end

you can access to each condition and its action using the condition as your key

julia> jl.map[:(x > 100)]
quote
    #= none:2 =#
    x + 1
end

similarly, we can easily construct a JLIfElse

jl = JLIfElse()
jl.map[:(x > 100)] = :(x + 1)
jl.map[:(x > 80)] = :(x + 2)
jl.otherwise = :(error("some error msg"))
jl

if >(x, 100)
    x + 1
elseif >(x, 80)
    x + 2
else
    error("some error msg")
end

now let's generate back to Expr so that we can give Julia back some executable expression

julia> codegen_ast(jl)
:(if x > 100
      x + 1
  elseif x > 80
      x + 2
  else
      error("some error msg")
  end)

You can find available syntax types in Syntax Types

Pattern Matching

Since Expronicon serves as the meta programming stdlib for MLStyle, you can also use the syntax types along with MLStyle, e.g

using MLStyle
using Expronicon

f = @λ begin
   JLFunction(;name=:foo, args) => (args, )
   JLFunction(;name=:boo, args) => (args, )
   _ => nothing
end

ex_foo = @expr function foo(x::Int, y::T) where {T <: Real}
    x + y
end

ex_boo = @expr function foo(x::Int)
    x
end

then we can check if our match function gives the right result

julia> f(ex_foo)
(Any[:(x::Int), :(y::T)],)

julia> f(ex_boo)
(Any[:(x::Int)],)

You can use any syntax types builtin as your expression template to match using MLStyle. If you define your own syntax type, you can also support pattern matching via @syntax_pattern.

Expronicon.@syntax_pattern — Macro

@syntax_pattern <syntax type> <syntax checker>

Example

struct MyFunction
    ex :: Expr
    test_field :: Any
end

is_xxx(ex::Expr) = Meta.isexpr(ex, :function)
is_xxx(_) = false

julia> MyFunction(ex::Expr) = MyFunction(ex, :aaa)
julia> @syntax_pattern(MyFunction, is_xxx)

julia> @match :(function f() end) begin
        MyFunction(;test_field) => test_field
    end
:aaa

Analysis Functions

Expronicon provides a lot common analysis functions, you can find the list of them in Analysis. you can use them to check if the expression satisfy certain property, e.g you can check if a given object is a struct definition via is_struct, or check if a given function definition supports keyword arguments via is_kw_function.

Transform Functions

You can find the list of them in Transform.

Transform functions usually takes an expression and returns an expression e.g sometimes you only want the name symbol of your function arguments

def = @expr JLFunction function foo(x::Int, y::Real=2)
end

julia> def.args
2-element Vector{Any}:
 :(x::Int)
 :($(Expr(:kw, :(y::Real), 2)))

julia> name_only.(def.args)
2-element Vector{Symbol}:
 :x
 :y

Code Generation Functions

The code generation functions help you generate other target expressions, e.g codegen_ast generates the Julia AST object Expr. All the syntax type can use codegen_ast to generate the corresponding Expr, there are also some other functions start with name codegen in CodeGen you may find useful.

Pretty Printing

Sometimes, when you define your own intermediate representation, you may want to pretty print your expression with colors and indents. Expronicon also provide some tools for this in Printings.