Analysis

Functions for analysing a given Julia Expr, e.g splitting Julia function/struct definitions etc.

Expronicon.JLFor — Method

JLFor(ex::Expr)

Create a JLFor from given Julia for loop expression.

Example

julia> ex = @expr for i in 1:10, j in 1:j
           M[i, j] += 1
       end
:(for i = 1:10, j = 1:j
      #= REPL[3]:2 =#
      M[i, j] += 1
  end)

julia> jl = JLFor(ex)
for i in 1 : 10,
    j in 1 : j
    #= loop body =#
    begin
        #= REPL[3]:2 =#        
        M[i, j] += 1        
    end
end

julia> jl.vars
2-element Vector{Any}:
 :i
 :j

julia> jl.iterators
2-element Vector{Any}:
 :(1:10)
 :(1:j)

Expronicon.JLFunction — Method

JLFunction(ex::Expr)

Create a JLFunction object from a Julia function Expr.

Example

julia> JLFunction(:(f(x) = 2))
f(x) = begin
    #= REPL[37]:1 =#    
    2    
end

Expronicon.JLIfElse — Method

JLIfElse(ex::Expr)

Create a JLIfElse from given Julia ifelse Expr.

Example

julia> ex = :(if foo(x)
             x = 1 + 1
         elseif goo(x)
             y = 1 + 2
         else
             error("abc")
         end)
:(if foo(x)
      #= REPL[41]:2 =#
      x = 1 + 1
  elseif #= REPL[41]:3 =# goo(x)
      #= REPL[41]:4 =#
      y = 1 + 2
  else
      #= REPL[41]:6 =#
      error("abc")
  end)

julia> JLIfElse(ex)
if foo(x)
    begin
        #= REPL[41]:2 =#        
        x = 1 + 1        
    end
elseif begin
    #= REPL[41]:3 =#    
    goo(x)    
end
    begin
        #= REPL[41]:4 =#        
        y = 1 + 2        
    end
else
    begin
        #= REPL[41]:6 =#        
        error("abc")        
    end
end

Expronicon.JLKwStruct — Type

JLKwStruct(ex::Expr, typealias=nothing)

Create a JLKwStruct from given Julia struct Expr, with an option to attach an alias to this type name.

Example

julia> JLKwStruct(:(struct Foo
           x::Int = 1
       end))
#= kw =# struct Foo
    #= REPL[39]:2 =#
    x::Int = 1
end

Expronicon.JLStruct — Method

JLStruct(ex::Expr)

Create a JLStruct object from a Julia struct Expr.

Example

julia> JLStruct(:(struct Foo
           x::Int
       end))
struct Foo
    #= REPL[38]:2 =#
    x::Int
end

Expronicon.compare_expr — Method

compare_expr(lhs, rhs)

Compare two expression of type Expr or Symbol semantically, which:

ignore the detail value LineNumberNode in comparision

ignore the detailed name of typevars in Expr(:curly, ...) or Expr(:where, ...)

This gives a way to compare two Julia expression semantically which means although some details of the expression is different but they should produce the same lowered code.

Expronicon.compare_vars — Method

compare_vars(lhs, rhs)

Compare two expression by assuming all Symbols are variables, thus their value doesn't matter, only where they are matters under this assumption. See also compare_expr.

Expronicon.has_kwfn_constructor — Function

has_kwfn_constructor(def[, name = struct_name_plain(def)])

Check if the struct definition contains keyword function constructor of name. The constructor name to check by default is the plain constructor which does not infer any type variables and requires user to input all type variables. See also struct_name_plain.

Expronicon.has_plain_constructor — Function

has_plain_constructor(def, name = struct_name_plain(def))

Check if the struct definition contains the plain constructor of name. By default the name is the inferable name struct_name_plain.

Example

def = @expr JLKwStruct struct Foo{T, N}
    x::Int
    y::N

    Foo{T, N}(x, y) where {T, N} = new{T, N}(x, y)
end

has_plain_constructor(def) # true

def = @expr JLKwStruct struct Foo{T, N}
    x::T
    y::N

    Foo(x, y) = new{typeof(x), typeof(y)}(x, y)
end

has_plain_constructor(def) # false

the arguments must have no type annotations.

def = @expr JLKwStruct struct Foo{T, N}
    x::T
    y::N

    Foo{T, N}(x::T, y::N) where {T, N} = new{T, N}(x, y)
end

has_plain_constructor(def) # false

Expronicon.is_field — Method

is_field(ex)

Check if ex is a valid field expression.

Expronicon.is_field_default — Method

is_field_default(ex)

Check if ex is a <field expr> = <default expr> expression.

Expronicon.is_for — Method

is_for(ex)

Check if ex is a for loop expression.

Expronicon.is_function — Method

is_function(def)

Check if given object is a function expression.

Expronicon.is_ifelse — Method

is_ifelse(ex)

Check if ex is an if ... elseif ... else ... end expression.

Expronicon.is_kw_function — Method

is_kw_function(def)

Check if a given function definition supports keyword arguments.

Expronicon.is_literal — Method

is_literal(x)

Check if x is a literal value.

Expronicon.is_struct — Method

is_struct(ex)

Check if ex is a struct expression.

Expronicon.is_struct_not_kw_struct — Method

is_struct_not_kw_struct(ex)

Check if ex is a struct expression excluding keyword struct syntax.

Expronicon.split_doc — Method

split_doc(ex::Expr) -> line, doc, expr

Split doc string from given expression.

Expronicon.split_function — Method

split_function(ex::Expr) -> head, call, body

Split function head declaration with function body.

Expronicon.split_function_head — Method

split_function_head(ex::Expr) -> name, args, kw, whereparams

Split function head to name, arguments, keyword arguments and where parameters.

Expronicon.split_struct — Method

split_struct(ex::Expr) -> ismutable, name, typevars, supertype, body

Split struct definition head and body.

Expronicon.split_struct_name — Method

split_struct_name(ex::Expr) -> name, typevars, supertype

Split the name, type parameters and supertype definition from struct declaration head.

Expronicon.support_default — Method

support_default(f)

Check if field type f supports default value.

Expronicon.@expr — Macro

@expr <type> <expression>

Return the expression in given type.

Example

julia> ex = @expr JLKwStruct struct Foo{N, T}
           x::T = 1
       end
#= kw =# struct Foo{N, T}
    #= /home/roger/code/julia/Expronicon/test/analysis.jl:5 =#
    x::T = 1
end

Expronicon.@expr — Macro

@expr <expression>

Return the original expression object.

Example

julia> ex = @expr x + 1
:(x + 1)

Expronicon.@test_expr — Macro

@test_expr <type> <ex>

Test if the syntax type generates the same expression ex. Returns the corresponding syntax type instance. Requires using Test before using this macro.

Example

def = @test_expr JLFunction function (x, y)
    return 2
end
@test is_kw_fn(def) == false

Expronicon.@test_expr — Macro

@test_expr <expr> == <expr>

Test if two expression is equivalent semantically, this uses compare_expr to decide if they are equivalent, ignores things such as LineNumberNode generated Symbol in Expr(:curly, ...) or Expr(:where, ...).