CodeGen

Code generators, functions that generates Julia Expr from given arguments, Expronicon types. You can import all the functions in this module by using Expronicon.CodeGen.

Expronicon.CodeGen — Module

collection of code generators.

Expronicon.CodeGen.codegen_ast_struct — Method

codegen_ast_struct(def)

Generate pure Julia struct Expr from struct definition. This is equivalent to codegen_ast for JLStruct. See also codegen_ast.

Example

julia> def = JLKwStruct(:(struct Foo
           x::Int=1
           
           Foo(x::Int) = new(x)
       end))
struct Foo
    x::Int = 1
end

julia> codegen_ast_struct(def)
:(struct Foo
      #= REPL[21]:2 =#
      x::Int
      Foo(x::Int) = begin
              #= REPL[21]:4 =#
              new(x)
          end
  end)

Expronicon.CodeGen.codegen_ast_struct_body — Method

codegen_ast_struct_body(def)

Generate the struct body.

Example

julia> def = JLStruct(:(struct Foo
           x::Int
           
           Foo(x::Int) = new(x)
       end))
struct Foo
    x::Int
end

julia> codegen_ast_struct_body(def)
quote
    #= REPL[15]:2 =#
    x::Int
    Foo(x::Int) = begin
            #= REPL[15]:4 =#
            new(x)
        end
end

Expronicon.CodeGen.codegen_ast_struct_curly — Method

codegen_ast_struct_curly(def)

Generate the struct name with curly if it is parameterized.

Example

julia> using Expronicon.Types, Expronicon.CodeGen

julia> def = JLStruct(:(struct Foo{T} end))
struct Foo{T}
end

julia> codegen_ast_struct_curly(def)
:(Foo{T})

Expronicon.CodeGen.codegen_ast_struct_head — Method

codegen_ast_struct_head(def)

Generate the struct head.

Example

julia> using Expronicon.Types, Expronicon.CodeGen

julia> def = JLStruct(:(struct Foo{T} end))
struct Foo{T}
end

julia> codegen_ast_struct_head(def)
:(Foo{T})

julia> def = JLStruct(:(struct Foo{T} <: AbstractArray end))
struct Foo{T} <: AbstractArray
end

julia> codegen_ast_struct_head(def)
:(Foo{T} <: AbstractArray)

Expronicon.CodeGen.codegen_match — Function

codegen_match(f, x[, line::LineNumberNode=LineNumberNode(0), mod::Module=Main])

Generate a zero dependency match expression using MLStyle code generator, the syntax is identical to MLStyle.

Example

codegen_match(:x) do
    quote
        1 => true
        2 => false
        _ => nothing
    end
end

This code generates the following corresponding MLStyle expression

@match x begin
    1 => true
    2 => false
    _ => nothing
end