Transform

Some common transformations for Julia Expr, these functions takes an Expr and returns an Expr.

Expronicon.eval_interp — Method

eval_interp(m::Module, ex)

evaluate the interpolation operator in ex inside given module m.

Expronicon.eval_literal — Method

eval_literal(m::Module, ex)

Evaluate the literal values and insert them back to the expression. The literal value can be checked via is_literal.

Expronicon.flatten_blocks — Method

flatten_blocks(ex)

Remove hierachical expression blocks.

Expronicon.name_only — Method

name_only(ex)

Remove everything else leaving just names, currently supports function calls, type with type variables, subtype operator <: and type annotation ::.

Example

julia> using Expronicon

julia> name_only(:(sin(2)))
:sin

julia> name_only(:(Foo{Int}))
:Foo

julia> name_only(:(Foo{Int} <: Real))
:Foo

julia> name_only(:(x::Int))
:x

Expronicon.prettify — Method

prettify(ex)

Prettify given expression, remove all LineNumberNode and extra code blocks.

Tips

the LineNumberNode inside macro calls won't be removed since the macrocall expression requires a LineNumberNode. See also issues/#9.

Expronicon.rm_annotations — Method

rm_annotations(x)

Remove type annotation of given expression.

Expronicon.rm_lineinfo — Method

rm_lineinfo(ex)

Remove LineNumberNode in a given expression.

Tips

the LineNumberNode inside macro calls won't be removed since the macrocall expression requires a LineNumberNode. See also issues/#9.

Expronicon.rm_nothing — Method

rm_nothing(ex)

Remove the constant value nothing in given expression ex.

Expronicon.subtitute — Method

subtitute(ex::Expr, old=>new)

Subtitute the old symbol old with new, similar to replace_symbol, but with a easier to read syntax.