Analysis

Functions for analysing a given Julia Expr, e.g splitting Julia function/struct definitions etc.

Expronicon.Analysis — Module

analysis functions for Julia Expr

Expronicon.Types.JLFunction — Method

JLFunction(ex::Expr)

Create a JLFunction object from a Julia function Expr.

Example

julia> JLFunction(:(f(x) = 2))
f(x) = begin
    #= REPL[37]:1 =#    
    2    
end

Expronicon.Types.JLIfElse — Method

JLIfElse(ex::Expr)

Create a JLIfElse from given Julia ifelse Expr.

Example

julia> ex = :(if foo(x)
             x = 1 + 1
         elseif goo(x)
             y = 1 + 2
         else
             error("abc")
         end)
:(if foo(x)
      #= REPL[41]:2 =#
      x = 1 + 1
  elseif #= REPL[41]:3 =# goo(x)
      #= REPL[41]:4 =#
      y = 1 + 2
  else
      #= REPL[41]:6 =#
      error("abc")
  end)

julia> JLIfElse(ex)
if foo(x)
    begin
        #= REPL[41]:2 =#        
        x = 1 + 1        
    end
elseif begin
    #= REPL[41]:3 =#    
    goo(x)    
end
    begin
        #= REPL[41]:4 =#        
        y = 1 + 2        
    end
else
    begin
        #= REPL[41]:6 =#        
        error("abc")        
    end
end

Expronicon.Types.JLKwStruct — Type

JLKwStruct(ex::Expr, typealias=nothing)

Create a JLKwStruct from given Julia struct Expr, with an option to attach an alias to this type name.

Example

julia> JLKwStruct(:(struct Foo
           x::Int = 1
       end))
#= kw =# struct Foo
    #= REPL[39]:2 =#
    x::Int = 1
end

Expronicon.Types.JLStruct — Method

JLStruct(ex::Expr)

Create a JLStruct object from a Julia struct Expr.

Example

julia> JLStruct(:(struct Foo
           x::Int
       end))
struct Foo
    #= REPL[38]:2 =#
    x::Int
end

Expronicon.Analysis.is_fn — Method

is_fn(def)

Check if given object is a function expression.

Expronicon.Analysis.is_kw_fn — Method

is_kw_fn(def)

Check if a given function definition supports keyword arguments.

Expronicon.Analysis.split_doc — Method

split_doc(ex::Expr) -> line, doc, expr

Split doc string from given expression.

Expronicon.Analysis.split_function — Method

split_function(ex::Expr) -> head, call, body

Split function head declaration with function body.

Expronicon.Analysis.split_function_head — Method

split_function_head(ex::Expr) -> name, args, kw, whereparams

Split function head to name, arguments, keyword arguments and where parameters.

Expronicon.Analysis.split_struct — Method

split_struct(ex::Expr) -> ismutable, name, typevars, supertype, body

Split struct definition head and body.

Expronicon.Analysis.split_struct_name — Method

split_struct_name(ex::Expr) -> name, typevars, supertype

Split the name, type parameters and supertype definition from struct declaration head.