Analysis

Functions for analysing a given Julia Expr, e.g splitting Julia function/struct definitions etc.

Expronicon.Analysis — Module

analysis functions for Julia Expr

Expronicon.Analysis.is_kw_fn — Method

is_kw_fn(def)

Check if a given function definition supports keyword arguments.

Expronicon.Analysis.split_call — Method

split_call(ex::Expr) -> name, args, kw, whereparams

Split call name, arguments, keyword arguments and where parameters.

Expronicon.Analysis.split_function — Method

split_function(ex::Expr) -> head, call, body

Split function head declaration with function body.

Expronicon.Analysis.split_struct — Method

split_struct(ex::Expr) -> ismutable, name, typevars, supertype, body

Split struct definition head and body.

Expronicon.Analysis.split_struct_name — Method

split_struct_name(ex::Expr) -> name, typevars, supertype

Split the name, type parameters and supertype definition from struct declaration head.