Transform

Some common transformations for Julia Expr, these functions takes an Expr and returns an Expr.

Expronicon.no_default — Constant

const no_default = NoDefault()

Constant instance for NoDefault that describes a field should have no default value.

Expronicon.prettify — Function

prettify(ex)

Prettify given expression, remove all LineNumberNode and extra code blocks.

Expronicon.rm_lineinfo — Function

rm_lineinfo(ex)

Remove LineNumberNode in a given expression.

Expronicon.flatten_blocks — Function

flatten_blocks(ex)

Remove hierachical expression blocks.

Expronicon.name_only — Function

name_only(ex)

Remove everything else leaving just names, currently supports function calls, type with type variables, subtype operator <: and type annotation ::.

Example

julia> using Expronicon

julia> name_only(:(sin(2)))
:sin

julia> name_only(:(Foo{Int}))
:Foo

julia> name_only(:(Foo{Int} <: Real))
:Foo

julia> name_only(:(x::Int))
:x

Expronicon.rm_annotations — Function

rm_annotations(x)

Remove type annotation of given expression.

Missing docstring.

Missing docstring for replace_symbol. Check Documenter's build log for details.

Missing docstring.

Missing docstring for subtitute. Check Documenter's build log for details.

Expronicon.eval_interp — Function

eval_interp(m::Module, ex)

evaluate the interpolation operator in ex inside given module m.

Missing docstring.

Missing docstring for eval_literal. Check Documenter's build log for details.