Types

Convenient types for storing analysis results of a given Julia Expr, or for creating certain Julia objects easily.

Expronicon.Types — Module

intermediate types for Julia expression objects.

Expronicon.Types.no_default — Constant

const no_default = NoDefault()

Constant instance for NoDefault that describes a field should have no default value.

Expronicon.Types.JLField — Type

JLField <: JLExpr
JLField(name, type, line)

Type describes a Julia field in a Julia struct.

Expronicon.Types.JLFunction — Type

JLFunction <: JLExpr

Type describes a Julia function declaration expression.

Expronicon.Types.JLKwField — Type

JLKwField <: JLExpr
JLKwField(name, type, line, default=no_default)

Type describes a Julia field that can have a default value in a Julia struct.

Expronicon.Types.JLKwStruct — Type

JLKwStruct <: JLExpr

Type describes a Julia struct that allows keyword definition of defaults.

Expronicon.Types.JLStruct — Type

JLStruct <: JLExpr

Type describes a Julia struct.

Expronicon.Types.NoDefault — Type

NoDefault

Type describes a field should have no default value.